Updated file v2 for task 0

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -8,7 +8,7 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (h == NULL)
 		return (EXIT_FAILURE);
